requirements.txt
# Core Deep Learning Libraries
tensorflow==2.13.0
keras==2.13.1

# Data Processing and Visualization
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
Pillow==10.0.0

# Machine Learning Utilities
scikit-learn==1.3.0

# Web Application Framework
streamlit==1.25.0

# Progress Bars and Utilities
tqdm==4.65.0

# Image Processing
albumentations==1.3.1

# Model Visualization
pydot==1.4.2
graphviz==0.20.1

# Jupyter Notebook (optional)
jupyter==1.0.0
ipykernel==6.25.0

# Additional Utilities
requests==2.31.0
zipfile36==0.1.3